+++ source/systemUtils.py
@@ -28,7 +28,6 @@ from typing_extensions import (
 	TypeVar,
 )
 
-from config import RegistryKey
 import winKernel
 import winreg
 import shellapi
@@ -195,6 +194,7 @@ def _isSystemClockSecondsVisible() -> bool:
 
 	@return: True if the 'ShowSecondsInSystemClock' value is 1, False otherwise.
 	"""
+	from config.registry import RegistryKey
 	registry_path = rf"{RegistryKey.CURRENT_VERSION.value}\Explorer\Advanced"
 	value_name = "ShowSecondsInSystemClock"
 	try: